Consolidated utils from src/util.dart into src/common/utils/utils.dart
