+++ lib/src/common/activity_service.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:github/src/common.dart';
 import 'package:github/src/common/model/users.dart';
 import 'package:github/src/common/util/pagination.dart';
-import 'package:github/src/util.dart';
+import 'package:github/src/common/util/utils.dart';
 import 'package:http/http.dart' as http;
 
 /// The [ActivityService] handles communication with activity related methods
