+++ lib/src/common/users_service.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:github/src/common.dart';
 import 'package:github/src/common/model/users.dart';
 import 'package:github/src/common/util/pagination.dart';
-import 'package:github/src/util.dart';
+import 'package:github/src/common/util/utils.dart';
 import 'package:http/http.dart' as http;
 
 /// The [UsersService] handles communication with user related methods of the
